Improve robustness of `verdi` in absence of profiles
It can occur, due to a problem during the first time setup of an AiiDA
instance, that the configuration file is left incomplete, i.e. without
any valid profiles or a default profile. This could render the `verdi`
command unusable as it assumed the configuration to always be complete.

We add a check to the `verdi` main command and add extract checks and
information to `verdi profile list` and `verdi profile show` in the case
of absent profiles or a default profile.