Add support for the export/import of `Log` entities
To make the export of `Log` entities possible, the `QueryBuilder` had to
be extended to support retrieving log records for a given `Node`. The
following changes were applied:

 * Internal `log_model_class` has been removed and replaced by `orm.Log`
 * Added support to join node to log (`with_node`)
 * Added support to join log to node (`with_log`)

The `verdi export create` has a new flag to include or exclude the
export of logs for nodes that are to be exported, defaulting to include:

 `--include-logs/--exclude-logs`

Documentation has been updated to include new `QueryBuilder` join args
in table and the `metadata.json` example has been updated in documentation
to include correct Log info.