Fix bug in `delete_nodes` when passing pks of non-existing nodes
The `aiida.manage.database.delete.nodes.delete_nodes` function was not
checking whether the pks that were passed correspond to existing nodes
which would cause the line that fetches the repository folders, after
the graph traversal queries had been completed, to except. To fix this
we first verify that the passed pks exist and if not print a warning and
discard them.