Fix upper limit for dependency to `pg8000<1.13.0` (#2456)
The library `pg8000` which is used by `pgtest` to interface with
Postgres and is used by us for testing purposed released a new minor
version `v1.13.0` on Feb 1 2019, in which it dropped `py2` support
causing our tests to fail miserably. For the time being, we put an upper
bound on that dependency.