Finalizing work of moving `Node` to new backend interface
A new abstract interface has been defined for `BackendNode` in the
module `aiida.orm.implementation.nodes`. This has subsequently been
implemented for the two current database backends Django and SqlAlchemy.
The role of the `BackendNode` class is to provide a single interface,
independent on the database backend implementation, to store and
retrieve data from `Node` instances in the database. The front end
`Node` class will then only ever interact with the backend entity that
it encapsulates and shall never directly interact with the database
model that is encapsulated within the `BackendNode` class.

This commit deprecates quite a few methods from the old node class,
notably all methods of the variant `*_db_*` that retrieved or set data
directly from the database. These are now part of the `BackendNode`
interface, but since all its methods will operate directly on the
database, the `db` infix has been dropped.

All business logic that does not operate directly on database data, such
as the caching of links and attributes of unstored nodes, is implemented
in the front end ORM class `aiida.orm.node.Node`. Also the biggest part
of the business logic is implemented here, to keep the backend layer
thin and lean.