Add exception for `calculation.job.JobCalculation` process type inference (#2428)
In the migration following the provenance redesign, the process type had
to be inferred from the existing type string for `JobCalculation` nodes.
The function `infer_calculation_entry_point` did not account for literal
`calculation.job.JobCalculation.` type strings being present and would
return `.JobCalculation` as the fallback process type. However, it makes
more sense to just have an empty process type string in this case, as
would be the case for a base `CalcJobNode` instance.