Catch `UnroutableError` for orphaned processes in `verdi process` calls (#2445)
If the RabbitMQ task of a running process is lost before it is
completed, it will never complete, but any `verdi process` commands that
try to reach it such as `pause`, `play` and `kill` will throw an
`UnroutableError`. This exception should be caught and an error message
should be printed instead.