Ensure nested `calcfunctions` work as long `store_provenance=False` (#2464)
In principle `calcfunctions` are not allowed to call other `calcfunctions`
because a calculation is not allowed to call other calculations.
However, `calcfunctions` can be called as normal functions by setting
the metadata variable `store_provenance` to `False`. Therefore, as long
as `store_provenance` is turned off for nested functions, a
`calcfunction` is allowed to call other `calcfunctions`.

To make this work, the `Process._setup_db_record` method had to be
adapted to only add a call link to the parent process, if and only if
`store_provenance` for the child node is enabled. Otherwise, the link
validation would fail since a `CALL` link outgoing of a
`CalculationNode` is forbidden.