Add exception in `load_node_class` for backwards compatibility
The `load_node_class` serves to turn a node type string into the
corresponding `Node` sub class. This is based on how those type strings
are constructed. Currently, they will start either with `data.` or
`process.` since the `node.` prefix has been dropped. However, some
recent database migrations targeted databases where the `node.` prefix
was still present and therefore so did the tests. However, when the
migration tests are run, they will do so with the current ORM, which no
longer considers type strings starting with `node.` to be valid. To make
the migration tests work without changing the migrations themselves,
which cannot be done, we need to add an exception in the node class
loader for type strings starting with `node.`. By just stripping it, it
should once again conform to the current node type spec and the logic
can be continued.