Restore the import speed of main package to keep `verdi` snappy (#2513)
* Make the `DBLogHandler` log configuration lazy upon loading database env

The logging needs to be configured upon loading the `aiida` package, but
we want to prevent the database environment or the `aiida.orm` module
from being loaded as this will slow down the initialization. This is
especially detrimental to the performance of `verdi` which everytime it
is loaded, i.e. also each time tab-completion is triggered, the whole
package needs to be loaded. To prevent this, the implementation of the
`DBLogHandler` was already moved from `aiida.common` to `aiida.orm.utils`
but since it is still being configured in `aiida.common.log`, the
logging module will import the `aiida.orm` module. To fix this, we
remove the handler from the default logging configuration and reinsert
it dynamically when the database environment is loaded, without which
the handler would be non-functional anyway.

* Copy implementation `plumpy` language tools instead of importing

The `aiida.common.lang` module will be imported when loading `verdi` and
importing the `protected` and `override` decorators from the `plumpy`
library were causing the whole library to be imported, slowing
everything down. By copying the implementation, the time for executing
`verdi` is roughly reduced by 25%.