Backport migration `7a6587e16f4c` of release `v0.12.3` (#2514)
Migration `7a6587e16f4c` was added to the `v0.12.3` release to improve the speed
of adding nodes to groups by adding a uniqueness constraint, with implied index,
on the `dbgroup_dbnodes` join table for SqlAlchemy. However, a similar change had
already been introduced in migration `59edaf8a8b79` to be released in `v1.0.0`
but in a slightly differently way. This would pose problems for people who would
migrate their database from `v0.12.3` to any version based on `v1.*`. Therefore,
special care needed to be taken to resolve this branching of revisisons.

To solve this branching in the migration revisions for the SqlAlchemy backend,
the new migration `7a6587e16f4c` is backported and added directly after revision
`35d4ee9a1b0e`, which is the point of divergence between `v0.12.*` and `v1.0.0`.
Revision `59edaf8a8b79` is then adapted to first check for the presence of the
changes introduced by the backported revision `7a6587e16f4c` and when found drop
them, before applying the changes it would have done originally. This method
guarantees that the migrations will work whether the database contains the
changes of the backported migration `7a6587e16f4c` or not.