Providing a non-ORM way to add nodes to a group. This speeds up a lot the group addition procedure and import method of SQLA should benefit from it (#2471)
Fast group addition, skipping the ORM

This PR solves issue #1319 and provides a fast way to add nodes to a group that circumvents the ORM. Import/export benefit from this; particularly import under SQLA.

In order to achieve this, it uses RAW SQL statements. It was agreed by @giovannipizzi  @sphuber 
 and @szoupanos that this is acceptable for aiida 0.X. sqlalchemy >= 1.1 provides better solutions for this in aiida 1.0 (see comments in the code).

Furthermore, it adds a new migration `7a6587e16f4c`.
For consistency, PR #2514 inserts this migration also in the upstream branch for aiida 1.0 (and handles collisions with other following migrations).