Fixing bug related to deleted nodes in recent log migration (#2507)
In a recent database migration, `0024_dblog_update` for Django and `041a79fc615f`
for SqlAlchemy, the `DbLog` table was updated where the `objpk` was replaced by
a proper foreign key to the `DbNode` row to which the log belongs. This required
the deletion of logs that did not correspond to node entities, which were typically
legacy workflows. The migration accounted for this by detecting these logs, exporting
them to a file for safekeeping, before deleting them. However, the scenario did
not consider that for the logs that remained, the corresponding node may have
been deleted. Since the relationship was not formalized through a foreign key this
would not have cascaded onto the related log entries. The migration would then
fail as a result when trying to the set the `dbnode_id` foreign key to the `ojbpk`
in the case that the node row no longer existed.

In this commit, the migrations are updated to detect orphaned log messages and
write them to a file before deleting them, just as is done for logs that belong
to legacy workflows or unknown entities.