Remove any use of our ORM in the migrations and tests
Many of the migrations and their respective unit tests used the ORM of
`aiida-core` or the database model classes directly. However, as the
database models evolve and change, during migrations one cannot rely on
the current implementation of the database models or the ORM as it may
not be commensurate of their state at the time of the migration. This
means that the migrations absolutely have to be written relying only on
raw SQL or on the model classes that are provided through the migration
mechanism of the various backends, as they will make sure that those
fake classes reflect the state of the models at the time of the migration.