Reorganize the `aiida.work` module (#2524)
The `aiida.work` module was conceived when the `WorkChain` class was
first implemented. Ever since, the `Process` class on which it is based
starting filling are more prominent role and slowly all running
processes were migrated to be a `Process` sub class. All of the logic
was added to the `aiida.work` module, which turned into the module that
implemented all the functionality of the engine of `aiida-core`.

Therefore we rename the `aiida.work` module to `aiida.engine` and
reorganize the internal structure to clarify the role of the different
components.