Formalizing importable resources on second level modules
The convention for those resources, e.g. classes and functions, that
should be usable by users of `aiida-core`, is that they should be
importable directly from a second level module in the hierarchy. That is
to say, anything that can be imported directly from `aiida.orm` is
intended to be used by users and therefore these resources come with the
guarantee that if they will be changed or moved, this will be done with
a decent deprecation path.