Remove the `FrozenDict` data sub class
The `FrozenDict` class was added to allow `ProcessNode` instances to be
passed to `Process` classes. However, its implementation is flawed as it
uses the pks of the nodes it wraps, which are not persistent across
databases and as such provenance will be lost when it is imported. We
remove the source code here. In a future commit, it will have to be
decided how to migrate existing data.