Reenable transport plugin tests
A few transport plugin tests had been commented out because due to a
change on Travis they started failing. The reason was that Travis added
files to the `.bashrc` that depended on the `TRAVIS_HOME` environment
variable to be set, which was not the case if one SSH'ed to localhost.
This caused the sourcing of the `.bashrc` to fail during the SSH
transport tests in which we SSH to localhost where this environment
variable is not set. As a workaround we add a step in the before install
section of the `.travis.yml` where we use `sed` to replace the
`TRAVIS_HOME` variable for `HOME` which will in fact be set.