Improve the efficiency of `DbLog` migration for SqlAlchemy (#2540)
The old implementation was using a `NOT IN` with the list of all node
ids to filter all log entries that had an `objpk` that did not
correspond to a node entry. This is extremely inefficient as it loads
the entire node pk column in memory and has to traverse it for each log
entry. In this commit, we change this to a construct that will simply
search for the `objpk` in the node table, which is indexed and simply
returns 1 if it exists, which is than checked in the filter of the logs.