Make various protected `Node` methods public
These methods are intended to be called from outside the class hierarchy
but were made protected (by prefixing them with an underscore) merely to
hide them from auto-completion and signal to users that they should not
be called by the user. However, this is the wrong usage of this feature
and methods that should be callable from the outside should not be
protected.