bug fix for `verdi node delete`
Fix a bug such that if no argument is passed to `verdi node delete`. It will try to delete all notes in the database.
This is because `QueryBuilder().append(Node, filters={'or': []}, project='id').all()` match to every single node and projects its id. 
Perhaps this behaviour of `QueryBuilder` is not as expected?