Add option to circumvent ORM in `Group.add_nodes` for SQLA (#2518)
This option is only available at the `BackendGroup` level and
therefore only to users who need this speedup (the ORM level
users should not be aware of such options). This was originally
added to speed up the import of archives for SQLA backends
whose implementation was modified to use this optimization.