Re-enable export functionality with exception of migrations. (#2554)
This re-enables the possibility to create export archives in the new format with
version `v0.4`. The migration of existing archives with an older version is not
yet implemented. Tests that rely on this functionality have been temporarily
disabled. The logic of the `verdi export migrate` is, however, already adapted
to support automatic recursive migration from any version to the latest.