Remove link type filter in `ProcessNode._get_objects_to_hash`
The `_get_objects_to_hash` methods called `get_incoming` with a link
type filter tat included all and the only allowed link types and so
the filter is redundant.

This commit also adds a `uuid` property to the `Process` class which
is a proxy to the UUID of the underlying `ProcessNode`.