Add an index to columns of `DbLink` for SqlAlchemy (#2561)
Indices are added for the `input_id`, `output_id` and `type` column of
the `DbLink` model for SqlAlchemy. This makes it congruent with the
Django model and ensures that the engine will not grind to a halt for
bigger production databases which often needs to get all incoming or
outgoing links for a node.