Prevent launching or construction of `WorkChain` and `CalcJob` base class (#2564)
The `WorkChain` and `CalcJob` classes are base classes that are meant to
be sub classed as they miss the necessary functionality to successfully
run. However, it was still possible to launch these base classes through
the launchers or straight up constructing them. They would immediately
fail when launched, but they would create a node in the provenance graph
nonetheless. To prevent this from happening, a check is placed in the
constructors to make sure that it is actually a sub class that is being
constructed.