Add efficient `count` method for `aiida.orm.groups.Group` (#2567)
Before the only way to get the number of entries in a `Group` was
through the iterator `nodes` and call `len()` on it. However, this is
almost always not the fastest way. Given that the count is used often in
for example `verdi group list -C` it is worthwhile to have a `count`
method that operates as direct on the database layer as possible.