Fix version comparison in import export
The export version number in newer `aiida-core` versions is no longer a
string but a float. Converting both to `StrictVersion` should solve the
problem. The import of `StrictVersion` from `distutils` triggered an
error in `pylint` and therefore `distutils` is currently ignored until
the issue is addressed: https://github.com/PyCQA/pylint/issues/73