Add unit tests for super call check in `Process.define`
Sub classes of `Process` such as `CalcJob` and `WorkChain` will
themselves be sub classed by users and they will have to implement the
`define` classmethod. It is crucial that the `super` is called here or
the instance will not work. To ensure this, there is a check in the
`Process.spec` class method, which verifies that the spec is set, which
indirectly means the `define` of the base class is called. This commit
adds two unit tests for `WorkChain` and `CalcJob` to make sure the
engine throws an `AssertionError` otherwise.