Make `Port.non_db` attribute inheritable from `PortNamespace`
If the `non_db` attribute is not explicitly defined, either upon
construction or through the setter, the value will inherited from the
namespace to which the port is added. This removes the requirement to
specify `non_db=True` for every port explicitly if the parent namespace
is already marked as non storable as well.