Update the documentation on processes in `aiida-core==1.0.0`
Since `aiida-core==1.0.0` everything that runs is a process and even
some earlier existing processes have changed their behavior and
implementation. The documentation is changed significantly to explain
these new concepts and the first parts of the new interface is
described. Some advanced sections remain to be written but issues have
been opened and a link to it is added in the corresponding sections
whose general structure has already been setup.