Fix version requirements for `plumpy` dependency
Dependency `plumpy` just released `v0.13.0` which makes it compatible
with `pika==1.0.0b2`, which is expected to be the last beta with
breaking changes before the real release, and with `kiwipy==0.5.0` which
also was just released to be compatible with `pika==1.0.0b2`.
Since `kiwipy` now introduces proper limits on its dependencies `pika`
and `topika`, we no longer have to those explicitly in `aiida-core` but
can simply rely on fixing the version of `plumpy`. The versions of the
four libraries that should work with this version of `aiida-core` are:

 * `plumpy==0.13.0`
 * `kiwipy==0.5.0`
 * `topika==0.2.0`
 * `pika==1.0.0b2`