Allow filelike in `SinglefileData.set_file` and constructor (#2638)
With the new node repository interface no longer providing access to
absolute filepaths, the requirement for a filepath in the
`SinglefileData` constructor and `set_file` method was too restrictive.
Here we add the possibility to add a filelike argument as well.

The `aiida.backends.tests.orm.data.test_data.py` file has been removed.
It was not added in the list of tests to be run and the tests it contained
were incorrect. The functionality they were intended to test is correctly
tested in `aiida.backends.tests.orm.node.test_node.py`.