Ensure `verdi import` returns non-zero status if incompatible archive version. (#2639)
Changed the message to critical if exception `IncompatibleArchiveVersionError`
is raised, which will automatically return a non-zero exit status. This
behavior is important for scripts that depend on this status to be able
to tell whether the command was successful. The check applies to both
local paths as well as URLS.

The type for the archive parameter has been changed to a new sub class
of `click.Path`, called `ImportPath` which first does the checks of its
base class `click.Path` followed by URL check using six's urllib.