Remove default valid types from the base process specification
The idea was that it would make it easy to define processes as
developers would not have to define valid types for the input and
output namespaces, but the reality is that instead it often hides subtle
bugs if users forget to define an explicit port or mistype an in- or
output. To prevent these suprises, the base process specification will
now have input and output namespace that are **not** dynamic.