Fix bug in `Parser.get_outputs_for_parsing` (#2668)
The method creates the mapping of outputs that will be passed to the
parse method by looping over the ports in the outputs namespace and
checking if their property `pass_to_parser` is set. Except, this
property is only defined for `CalcJobOutputPorts` but the namespace can
also contain normal `PortNamespaces` which would cause an exception.
An explicit check is added to prevent this from happening.

Additionally, the `process_class` property was essentially redefining
the `load_process_class` of its base `Process` class, but did not allow
a fallback loading mechanism for process types that were not based on an
entry point string. The `CalcJob.process_class` has been removed and the
`Process.load_process_class` method has been turned into a property and
renamed to `process_class`. From the point of view of the `CalcJob`
nothing changes, except that now it *is* possible to load the process
class if it is a local module path and not an entry point string.