Add the `-c/--clear` option to `verdi group delete`
Before the `--force` flag was used to both disable asking for
confirmation as well as triggering to remove all nodes before deleting
the group. This bundling of two very different behaviors in a single
important flag is dangerous. Here we change `--force` to only deal with
asking for confirmation and add the `--clear` flag to allow the user to
confirm they are sure they want to delete a group that contains nodes.