Fix bug in `verdi group description` and `verdi group rename`
The doc string of `verdi group description` said that the description
would be echoed if no argument was specified but the argument was
actually required. The argument is now optional and when not provided
will trigger the current description of the group to be echoed.

The command `verdi group rename` is renamed to `verdi group relabel`
since a while back the `name` property was renamed to `label`.