Warn user when copying group to non-empty destination group
Fixes issue #2679.

Logic:
If destination group was not created and is not empty,
the user will be warned and asked to confirm the copy.

The copy will take place - or the user will have aborted.

Added test:
`test_copy_existing_group`: Creates source group with nodes.
First copies and creates destination group, checking all is well,
then copies the same source group to the same destination group,
checking an abort error is raised and the destinations group is unchanged.

Minor things:
Corrected typo in doc-string for `__init__` of Group entity.