Make the outputs of a live process nestable
The outputs namespace of a process specification allows for arbitrary
nesting, just like the inputs. The inputs of a process, when passed,
will have to have a nesting structure that matches that of the inputs
process specification. However, the internal structure of outputs of a
process were not keeping the same nested structure as their process
specification counterpart. Of course when the processes finishes and the
outputs have to be linked up through links to the process node, the
support for nesting disappears and the mapping will have to be
flattened. However, while the process is "alive" the internal outputs
should keep the nesting structure of the process spec.