Formalize the mapping of port namespaces to link labels
The `Process` and `ProcessNode` duality forces us to define a mapping
between the nested port namespaces of the process and the flat link
label hierarchy of the corresponding node in the provenance graph.

Originally the nested namespace were imploded into a single string,
using a single underscore to concatenate the various port names. This
made it however impossible to distinguish a link label that came from a
nested namespace or simply from a single port that contained an
underscore. Given that underscores are allowed in valid python
identifiers we don't want to forbid the use of those. Instead, we
change the nesting character to be a double underscore and forbid a port
name if it satisfies any of the following rules:

 * Contains more than one consecutive underscore
 * Starts with an underscore
 * Ends with an underscore
 * Is otherwise not a valid python identifier

With these restrictions, mapping the flattened link labels back onto a
nested namespaced dictionary becomes an injective function.