Make `Parser` an abstract class and `parse` abstract method
This prevents one from instantiating a parser that has not implemented
the `parse` method, without which it will be useless. This also makes
the validation of the `CalcJob.metadata.options.parser_name` easier as
it will now be able to verify that the specified parser entry point can
both be resolved and imported, as well as constructed, or the input
validation will fail. This will prevent users from accidentally
launching a calculation job only to find after it has finished that an
incorrect parser was specified.

Note that we cannot use the `validator` key of the input port for this
task, as the ports are overridable by sub classes of the `CalcJob` and
if they forget to reattach the validator, the validation will not be
performed.