Implement a generic `ProcessNode.get_builder_restart`
This new method will be able to return an instance of the `ProcessBuilder`
class that is already configured to run the same process that created
the process node on which it is called. It also prepopulates the builder
with the inputs of the process node. This makes it trivial to relaunch a
process with slightly different inputs starting from the original node.

To facilitate this functionality, the `ProcessBuilderNamespace` has been
changed from a normal `Mapping` to a `MutableMapping`. This does block
certain mapping methods, e.g. `values()` and `items()` if those also
happen to be the names of ports in the process specification but if the
port value setters had to be namespaced, e.g. `inputs`, the use of
namespaced ports would become very ugly:

  `builder.inputs.name.inputs.space.inputs.value = 1

instead of

  `builder.name.space.value = 1`

Instead, now one can simply cast a builder to a dict and all the mapping
methods will be once again available:

  `dict(builder).values()`

The `update` method is very important for the `get_builder_restart` and
so it has been implemented as `_update` in order to not block a process
from having a port with the name `update`.

Finally, the ad-hoc implementation of `CalcJobNode.get_builder_restart`
has been simplified by calling the parent and then simply setting the
options in the metadata namespace.