Fix duplicate flag in `verdi node delete`
The `-n` flag was used both for the dry run option as well as the force
option. To add the force option, the `NON_INTERACTIVE` predefined was
used which uses the `-n` flag. However, the name "force" makes more
sense in this case and since it is a dangerous flag, we opt to only
provide the long form `--force`.