Remove the legacy workflow tables (#2703)
The migrations that will drop the tables will first check if there is
any data in the tables, in which the user will be prompted with a
warning message and a request to continue or abort. Should the user
decide to continue, the content of the tables will be serialized to JSON
format and written to a file.