Change the heuristics of the `AIIDA_PATH` environment variable
The location of the configuration folder will be determined and
optionally created following these heuristics:

 * If the `AIIDA_PATH` variable is set, all the paths will be checked
   to see if they contain a configuration folder. The first one to be
   encountered will be set as `AIIDA_CONFIG_FOLDER`. If none of them
   contain one, a configuration folder will be created in the last
   path considered.

 * If the `AIIDA_PATH` variable is not set the `DEFAULT_AIIDA_PATH`
   value will be used as base path and if it does not yet contain a
   configuration folder, one will be created.

This new approach will guarantee that a configuration folder will
always be set and created if necessary. The one downside is that if
someone sets the environment variable to the wrong value, a config
folder will be created without a warning.