Add validation of the link label in `Node.add_incoming`
Because the link label is often used as a python identifier, for example
in the graph traversal tools, it is important that they are in fact
valid python identifiers. To guarantee this validation is added to the
`Node.add_incoming` method on the link label value. Valid link labels
adhere to the following restrictions:

 * Has to be a valid python identifier
 * Can only contain alphanumeric characters and underscores
 * Can not start or end with an underscore

Validation of port names will call the same validation function, since
the port name will become (part) of the eventual link name. An
additional constraint for port names is that they cannot contain more
than one consecutive underscores, since a double underscore is used to
indicate a nested namespacing.