Implement generic list print function for command line
Multiple `verdi` commands will print a simple list of entries in a
collection where optionally some entries are highlighted. This concept
is abstracted in `aiida.cmdline.utils.echo.echo_formatted_list` to make
the output across the commands more homogeneous. The command that
implement it for now:

 * `verdi computer list`
 * `verdi profile list`
 * `verdi user list`

The function provides optional lambdas that can help sort the
collection, determine highlighting and which entries should be hidden.