Automatically add optional endopts marker to `verdi` commands (#2720)
Some `verdi` commands have options with the `MultipleValueOption` class
which allow to pass multi values to an option without repeating the flag.
However, this introduces the traditional ambiguity about when the option
values stop and become for example a positional argument, since the
option matching is greedy. The endopts marker `--` is traditionally used
to signal to the parser that the options have ended and everything that
follows should be interpreted as positional arguments.

To reduce the confusion of this feature, it would help if the usage
string of command with multi value options would include the endopts
marker. Here we override the `collect_usage_pieces` of the
`click.Command` class, which will ensure a usage string like e.g.:

    verdi export create [OPTIONS] [--] OUTPUT_FILE

The added `[--]` indicate that the endopts can optionally be used.