Implement `ConfigFileOption` reusable command line option type
This option type allows to add an option to command line interfaces that
will allow all its other options to be set through a yaml configuration
file that is passed to the config option. This functionality requires
the added click plugin dependency `click-config-file`. The new functionality
is implemented for `verdi quicksetup` and `verdi setup`.